DB: handle very large integers
The job_id values already exceed signed 32-bit integer space, which is
what postgresql uses when the Integer type is chosen. We hadn't noticed
it earlier because sqlite (which we were using for tests) treats Integer
as a 64-bit signed integer type.

This makes all of the identifiers explicitly BigInteger (64-bit or
larger) to handle future compatibility. There should be no meaningful
performance impact unless someone is running on a 32-bit kernel.

+++ logdetective/server/database/models/merge_request_jobs.py
@@ -7,7 +7,7 @@ import backoff
 from sqlalchemy import (
     Enum,
     Column,
-    Integer,
+    BigInteger,
     DateTime,
     String,
     ForeignKey,
@@ -33,22 +33,22 @@ class GitlabMergeRequestJobs(Base):
 
     __tablename__ = "gitlab_merge_request_jobs"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     forge = Column(Enum(Forge), nullable=False, index=True, comment="The forge name")
     project_id = Column(
-        Integer,
+        BigInteger,
         nullable=False,
         index=True,
         comment="The project gitlab id",
     )
     mr_iid = Column(
-        Integer,
+        BigInteger,
         nullable=False,
         index=False,
         comment="The merge request gitlab iid",
     )
     job_id = Column(
-        Integer,
+        BigInteger,
         nullable=False,
         index=True,
         comment="The job gitlab id",
@@ -161,9 +161,9 @@ class Comments(Base):
 
     __tablename__ = "comments"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     merge_request_job_id = Column(
-        Integer,
+        BigInteger,
         ForeignKey("gitlab_merge_request_jobs.id"),
         nullable=False,
         unique=True,  # 1 comment for 1 job
@@ -329,9 +329,9 @@ class Reactions(Base):
 
     __tablename__ = "reactions"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     comment_id = Column(
-        Integer,
+        BigInteger,
         ForeignKey("comments.id"),
         nullable=False,
         index=True,
@@ -343,7 +343,7 @@ class Reactions(Base):
         comment="The type of reaction",
     )
     count = Column(
-        Integer,
+        BigInteger,
         nullable=False,
         comment="The number of reactions, of this type, given in the comment",
     )
